tuan commit

+++ cinema-system/src/app/Components/component-tuan/list-discount/list-discount.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import {FilmService} from "../service/film.service";
 import {Promotions} from "../model/Promotions";
+import {FormBuilder} from "@angular/forms";
 
 @Component({
   selector: 'app-list-discount',
@@ -9,37 +10,79 @@ import {Promotions} from "../model/Promotions";
 })
 export class ListDiscountComponent implements OnInit {
   public discount: Array<Promotions> = [];
-  public editField: string;
-  constructor(private discountService: FilmService) { }
+  public discountCreate: Array<Promotions> = [];
+  public discountEdit: Array<Promotions> = [];
+  public page = 1;
+
+  // public editField: string;
+  // isHidden: boolean = true;
+
+
+  constructor(private discountService: FilmService) {
+  }
 
   ngOnInit(): void {
     this.discountService.getAllPromotion().subscribe(data => {
       this.discount = data;
+      this.discountCreate = [];
+      this.discountEdit = [];
+      this.discount.forEach((item, index) => {
+        this.discount[index].isEdit = false;
+      })
     });
   }
-
-  delete(id) {
-    this.discountService.deletePromotion(id).subscribe(result => {
-      this.ngOnInit();
-    }, error => console.error(error))
+  delete(id, index) {
+    if (id == null) {
+      this.discount.splice(index, 1);
+      console.log(this.discountCreate);
+    } else {
+      if(confirm("Bạn có chắc muốn xóa Khuyến mãi này không???")){
+        this.discountService.deletePromotion(id).subscribe(result => {
+          this.ngOnInit();
+        }, error => console.error(error));
+      }
+    }
   }
 
-  addRow(index) {
-    this.discountService.createPromotion(this.discount[index]).subscribe();
-    return true;
-  }
 
   addNew() {
-    this.discount.push(new Promotions());
+    let promo = new Promotions();
+    this.discount.push(promo);
+    promo.isEdit = true;
+    console.log(this.discount.length);
   }
 
-  edit(index){
-    this.discountService.editPromotion(this.discount[index], index+1).subscribe();
-    return true;
+  edit(id,index) {
+    if (id !== null && id !== undefined){
+      this.discountEdit.push(this.discount[index]);
+    }
+    this.discount[index].isEdit = !this.discount[index].isEdit;
   }
 
-  // changeValue(id: number, property: string, event: any) {
-  //   this.editField = event.target.textContent;
-  // }
+
+  saveAll() {
+    if (this.discountEdit != []) {
+      for (let i = 0; i < this.discountEdit.length; i++) {
+        if (this.discount[i].id == null) {
+          this.discount.splice(i, 1);
+        }
+      }
+      this.discountService.editAllPromotion(this.discountEdit).subscribe(data => {
+      });
+    }
+
+    for (let i = 0; i < this.discount.length; i++) {
+      if (this.discount[i].id == null) {
+        this.discountCreate.push(this.discount[i]);
+      }
+    }
+    if (this.discountCreate != []) {
+      console.log(this.discount);
+      this.discountService.createAllPromotion(this.discountCreate).subscribe(data => {
+      });
+    }
+    location.reload();
+  }
+
 
 }
